refactor: dispatch slack org callbacks through ccstate
