+++ turbo/apps/api/src/signals/routes/__tests__/integrations.bdd.test.ts
@@ -1,9 +1,12 @@
 import { createHash, createHmac, randomInt, randomUUID } from "node:crypto";
 
 import { OFFICIAL_TELEGRAM_BOT_ID } from "@vm0/api-contracts/contracts/zero-integrations-telegram";
+import { agentRunCallbacks } from "@vm0/db/schema/agent-run-callback";
+import { eq } from "drizzle-orm";
 import { HttpResponse, http } from "msw";
 import { describe, expect, it } from "vitest";
 
+import { db } from "../../../lib/db";
 import { testContext } from "../../../__tests__/test-helpers";
 import { env, mockEnv, mockOptionalEnv } from "../../../lib/env";
 import { now } from "../../../lib/time";
@@ -14,7 +17,6 @@ import {
   agentPhoneBddWebhookSecret,
   createBddIntegrationApi,
   telegramLoginAuth,
-  type ForwardedInternalCallback,
 } from "./helpers/api-bdd-integrations";
 import { createRunsAutomationsApi } from "./helpers/api-bdd-runs-automations";
 import { createWebhookCallbackApi } from "./helpers/api-bdd-webhooks";
@@ -163,14 +165,6 @@ function slackPostMessageCallsJson(): string {
   return JSON.stringify(context.mocks.slack.chat.postMessage.mock.calls);
 }
 
-function slackOrgCallbackDeliveries(
-  log: readonly ForwardedInternalCallback[],
-): readonly ForwardedInternalCallback[] {
-  return log.filter((entry) => {
-    return entry.path === "/api/internal/callbacks/slack/org";
-  });
-}
-
 async function waitForExpectation(assertion: () => void): Promise<void> {
   await expect
     .poll(async () => {
@@ -180,13 +174,39 @@ async function waitForExpectation(assertion: () => void): Promise<void> {
     .toBe(true);
 }
 
-async function waitForSlackOrgCallback(
-  log: readonly ForwardedInternalCallback[],
-  expected: ForwardedInternalCallback,
+async function readSlackOrgCallback(runId: string) {
+  const [callback] = await db()
+    .select({
+      url: agentRunCallbacks.url,
+      internalKind: agentRunCallbacks.internalKind,
+      status: agentRunCallbacks.status,
+      attempts: agentRunCallbacks.attempts,
+      deliveredAt: agentRunCallbacks.deliveredAt,
+      lastError: agentRunCallbacks.lastError,
+    })
+    .from(agentRunCallbacks)
+    .where(eq(agentRunCallbacks.runId, runId))
+    .limit(1);
+  return callback;
+}
+
+async function waitForSlackOrgCallbackRow(
+  runId: string,
+  expected: {
+    readonly status: "pending" | "delivered" | "failed";
+    readonly attempts?: number;
+    readonly lastError?: string | null;
+  },
 ): Promise<void> {
-  await waitForExpectation(() => {
-    expect(slackOrgCallbackDeliveries(log).at(-1)).toStrictEqual(expected);
-  });
+  await expect
+    .poll(async () => {
+      return await readSlackOrgCallback(runId);
+    })
+    .toMatchObject({
+      internalKind: "slack:org",
+      url: null,
+      ...expected,
+    });
 }
 
 async function pollRunnerRun(
@@ -1133,7 +1153,6 @@ describe("INT-01: Slack app deep webhook flows", () => {
     runs.acceptTelemetryIngest();
     integrations.configureSlackAppMocks();
     integrations.acceptSlackSessionHistoryDownloads();
-    integrations.forwardSlackInternalCallbacks();
     const runnerGroup = runs.configureRunnerGroup();
     await runs.grantProEntitlement(actor);
     await integrations.configureSlackRunModelPolicies(actor);
@@ -2269,7 +2288,6 @@ describe("INT-01: Slack app deep webhook flows", () => {
     runs.acceptTelemetryIngest();
     integrations.configureSlackAppMocks();
     integrations.acceptSlackSessionHistoryDownloads();
-    const callbackLog = integrations.forwardSlackInternalCallbacks();
     const runnerGroup = runs.configureRunnerGroup();
     await runs.grantProEntitlement(actor);
     await integrations.configureSlackRunModelPolicies(actor);
@@ -2291,6 +2309,11 @@ describe("INT-01: Slack app deep webhook flows", () => {
       channel: channelId,
     });
     const run1Id = await pollSlackRun(runnerGroup);
+    await waitForSlackOrgCallbackRow(run1Id, {
+      status: "pending",
+      attempts: 0,
+      lastError: null,
+    });
     const claim1 = await runs.claimRunnerJob(run1Id);
     context.mocks.slack.assistant.threads.setStatus.mockClear();
     await webhooks.requestAgentHeartbeat(
@@ -2307,10 +2330,10 @@ describe("INT-01: Slack app deep webhook flows", () => {
         status: "is thinking...",
       });
     });
-    await waitForSlackOrgCallback(callbackLog, {
-      path: "/api/internal/callbacks/slack/org",
-      status: 200,
-      body: { success: true },
+    await waitForSlackOrgCallbackRow(run1Id, {
+      status: "pending",
+      attempts: 0,
+      lastError: null,
     });
 
     integrations.mockSlackRunResultOutput("SLACK_BDD_OUTPUT");
@@ -2345,6 +2368,11 @@ describe("INT-01: Slack app deep webhook flows", () => {
     });
     const run1 = await runs.readRun(actor, run1Id);
     expect(run1.status).toBe("completed");
+    await waitForSlackOrgCallbackRow(run1Id, {
+      status: "delivered",
+      attempts: 1,
+      lastError: null,
+    });
 
     await integrations.postSlackEvent(teamId, {
       type: "app_mention",
@@ -2429,10 +2457,10 @@ describe("INT-01: Slack app deep webhook flows", () => {
         }),
       );
     });
-    await waitForSlackOrgCallback(callbackLog, {
-      path: "/api/internal/callbacks/slack/org",
-      status: 200,
-      body: { success: true },
+    await waitForSlackOrgCallbackRow(run4Id, {
+      status: "delivered",
+      attempts: 1,
+      lastError: null,
     });
     const run4 = await runs.readRun(actor, run4Id);
     expect(run4.status).toBe("failed");
@@ -2482,10 +2510,10 @@ describe("INT-01: Slack app deep webhook flows", () => {
       sandboxToken: claim6.sandboxToken,
       cliAgentType: "claude-code",
     });
-    await waitForSlackOrgCallback(callbackLog, {
-      path: "/api/internal/callbacks/slack/org",
-      status: 400,
-      body: { error: "Slack API error: channel_not_found" },
+    await waitForSlackOrgCallbackRow(run6Id, {
+      status: "failed",
+      attempts: 1,
+      lastError: "Slack API error: channel_not_found",
     });
     const run6 = await runs.readRun(actor, run6Id);
     expect(run6.status).toBe("completed");
@@ -2497,7 +2525,6 @@ describe("INT-01: Slack app deep webhook flows", () => {
     runs.acceptTelemetryIngest();
     integrations.configureSlackAppMocks();
     integrations.acceptSlackSessionHistoryDownloads();
-    const callbackLog = integrations.forwardSlackInternalCallbacks();
     const runnerGroup = runs.configureRunnerGroup();
     await runs.grantProEntitlement(actor);
     await bdd.readOnboardingStatus(actor);
@@ -2533,10 +2560,10 @@ describe("INT-01: Slack app deep webhook flows", () => {
       { authorization: `Bearer ${claim1.sandboxToken}` },
       [200],
     );
-    await waitForSlackOrgCallback(callbackLog, {
-      path: "/api/internal/callbacks/slack/org",
-      status: 200,
-      body: { success: true },
+    await waitForSlackOrgCallbackRow(run1Id, {
+      status: "pending",
+      attempts: 0,
+      lastError: null,
     });
 
     integrations.mockSlackRunResultOutput("NO_MODEL_FOOTER_OUTPUT");
@@ -2578,10 +2605,10 @@ describe("INT-01: Slack app deep webhook flows", () => {
       { authorization: `Bearer ${claim2.sandboxToken}` },
       [200],
     );
-    await waitForSlackOrgCallback(callbackLog, {
-      path: "/api/internal/callbacks/slack/org",
-      status: 200,
-      body: { success: true },
+    await waitForSlackOrgCallbackRow(run2Id, {
+      status: "pending",
+      attempts: 0,
+      lastError: null,
     });
     expect(
       context.mocks.slack.assistant.threads.setStatus,
@@ -2593,10 +2620,10 @@ describe("INT-01: Slack app deep webhook flows", () => {
       sandboxToken: claim2.sandboxToken,
       cliAgentType: "claude-code",
     });
-    await waitForSlackOrgCallback(callbackLog, {
-      path: "/api/internal/callbacks/slack/org",
-      status: 404,
-      body: { error: "Slack installation not found" },
+    await waitForSlackOrgCallbackRow(run2Id, {
+      status: "failed",
+      attempts: 1,
+      lastError: "Slack installation not found",
     });
     expect(context.mocks.slack.chat.postMessage).not.toHaveBeenCalled();
     const run2 = await runs.readRun(actor, run2Id);
